#tinygpmp: edits to `stm.c` with working 3 pin control
