+++ tinygpmp/Makefile_STM
@@ -1,42 +1,10 @@
-# Makefile
-
-TARGET = main
-SRC = stm.c
-
-# Adjust these settings based on your STM32 model
-STM32_MODEL = STM32F4xx
-
-# Compiler and linker options
-CC		= arm-none-eabi-gcc
-OBJCOPY	= arm-none-eabi-objcopy
-LD 		= arm-none-eabi-ld
-CFLAGS 	= -mlittle-endian -mthumb -mcpu=cortex-m4 -mthumb-interwork -mfloat-abi=hard -mfpu=fpv4-sp-d16
-LDFLAGS = -T linker.ld -specs=nosys.specs
-
-# Object files
-OBJ = $(SRC:.c=.o)
-
-# Build rules
-
-all: $(TARGET).elf $(TARGET).hex $(TARGET).bin
-
-$(TARGET).elf: $(OBJ)
-	$(CC) $(CFLAGS) $(LDFLAGS) -o $@ $^
-
-%.hex: %.elf
-	$(OBJCOPY) -O ihex $< $@
-
-%.bin: %.elf
-	$(OBJCOPY) -O binary $< $@
-
-%.o: %.c
-	$(CC) $(CFLAGS) -c $< -o $@
+all:
+	arm-none-eabi-gcc -ggdb -Wall BLINK_WKG2.c -T linker_script.ld -o blink.elf -mcpu=cortex-m3 -mthumb -nostdlib
+	arm-none-eabi-objcopy -O binary blink.elf blink.bin
 
 flash:
-	st-flash write main.bin 0x8000000
-
-.PHONY: clean
+	st-flash write blink.bin 0x08000000
 
 clean:
-	rm -f $(OBJ) $(TARGET).elf $(TARGET).hex $(TARGET).bin
+	rm -f *.bin *.hex *.elf
